requirements.txt
asgiref           3.7.2
Django            4.2.5
et-xmlfile        1.1.0
openpyxl          3.1.2
pip               23.2.1
python-dotenv     1.0.0
setuptools        68.1.2
sqlparse          0.4.4
typing_extensions 4.8.0
wheel             0.41.2
